Extract SquadService from ShowSquad View to follow thin-orchestrator pattern
Move all business logic (player enrichment, KPI calculations, depth chart,
contract watchlist, alerts, renewal data) from ShowSquad View into a new
SquadService::buildSquadOverview() method. The View now only loads the game,
calls the service, and returns the Blade view.
